strip down database/bip44
