+++ app/api/ada/lib/storage/database/wallet/tables.js
@@ -2,6 +2,10 @@
 
 import { Type } from 'lovefield';
 import type { lf$schema$Builder } from 'lovefield';
+import { KeyDerivationSchema } from '../primitives/tables';
+
+// TODO: get rid of this import
+import { Bip44WrapperSchema }  from '../bip44/tables';
 
 export type ConceptualWalletInsert = {|
   CoinType: number,
@@ -21,6 +25,30 @@ export const ConceptualWalletSchema: {
     ConceptualWalletId: 'ConceptualWalletId',
     CoinType: 'CoinType',
     Name: 'Name',
+  }
+};
+
+export type PublicDeriverInsert = {|
+  Bip44WrapperId: number,
+  KeyDerivationId: number,
+  Name: string,
+  LastSyncInfoId: number,
+|};
+export type PublicDeriverRow = {|
+  PublicDeriverId: number, // serial
+  ...PublicDeriverInsert,
+|};
+export const PublicDeriverSchema: {
+  +name: 'PublicDeriver',
+  properties: $ObjMapi<PublicDeriverRow, ToSchemaProp>
+} = {
+  name: 'PublicDeriver',
+  properties: {
+    PublicDeriverId: 'PublicDeriverId',
+    Bip44WrapperId: 'Bip44WrapperId',
+    KeyDerivationId: 'KeyDerivationId',
+    Name: 'Name',
+    LastSyncInfoId: 'LastSyncInfoId',
   }
 };
 
@@ -103,6 +131,30 @@ export const populateWalletDb = (schemaBuilder: lf$schema$Builder) => {
       ([ConceptualWalletSchema.properties.ConceptualWalletId]: Array<string>),
       true,
     );
+
+  // PublicDeriver
+  schemaBuilder.createTable(PublicDeriverSchema.name)
+    .addColumn(PublicDeriverSchema.properties.PublicDeriverId, Type.INTEGER)
+    .addColumn(PublicDeriverSchema.properties.Bip44WrapperId, Type.INTEGER)
+    .addColumn(PublicDeriverSchema.properties.KeyDerivationId, Type.INTEGER)
+    .addColumn(PublicDeriverSchema.properties.Name, Type.STRING)
+    .addColumn(PublicDeriverSchema.properties.LastSyncInfoId, Type.INTEGER)
+    .addPrimaryKey(
+      ([PublicDeriverSchema.properties.PublicDeriverId]: Array<string>),
+      true
+    )
+    .addForeignKey('PublicDeriver_Bip44Derivation', {
+      local: PublicDeriverSchema.properties.KeyDerivationId,
+      ref: `${KeyDerivationSchema.name}.${KeyDerivationSchema.properties.KeyDerivationId}`
+    })
+    .addForeignKey('PublicDeriver_Bip44Wrapper', {
+      local: PublicDeriverSchema.properties.Bip44WrapperId,
+      ref: `${Bip44WrapperSchema.name}.${Bip44WrapperSchema.properties.Bip44WrapperId}`
+    })
+    .addForeignKey('PublicDeriver_LastSyncInfo', {
+      local: PublicDeriverSchema.properties.LastSyncInfoId,
+      ref: `${LastSyncInfoSchema.name}.${LastSyncInfoSchema.properties.LastSyncInfoId}`
+    });
 
   // LastSyncInfoSchema Table
   schemaBuilder.createTable(LastSyncInfoSchema.name)
